add config functions to configs.__init__ and refactor
