+++ tests/torch_datasets/test_pvnet_uk_regional.py
@@ -2,8 +2,7 @@ import pytest
 import tempfile
 
 from ocf_data_sampler.torch_datasets.pvnet_uk_regional import PVNetUKRegionalDataset
-from ocf_data_sampler.config.load import load_yaml_configuration
-from ocf_data_sampler.config.save import save_yaml_configuration
+from ocf_data_sampler.config import load_yaml_configuration, save_yaml_configuration
 from ocf_datapipes.batch import BatchKey, NWPBatchKey
